@@ -1,8 +1,9 @@
 <Project>
 	<PropertyGroup>
-		<Version>1.0.0</Version>
 		<!--This will get replaced by Build System-->
-		<LangVersion>10.0</LangVersion>
+		<Version>1.0.0</Version>
+		
+		<LangVersion>11.0</LangVersion>
 		
 		<!--TODO <Nullable>enable</Nullable>-->
 
