@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFrameworks>net6.0;netcoreapp3.1;net48;net461</TargetFrameworks>
+		<TargetFrameworks>net7.0;net6.0;netcoreapp3.1;net48;net461</TargetFrameworks>
 		
 	</PropertyGroup>
 
